Update group manage for case-insensitive user-adding, organized components into folders
