+++ happiness-backend/config.py
@@ -18,7 +18,7 @@ class Config:
     # Security measures
     SECURITY_PASSWORD_SALT = os.environ.get("SECURITY_PASSWORD_SALT")
     SECRET_KEY = os.environ.get("SECRET_KEY")
-    SERVER_NAME = "happiness-app-backend.herokuapp.com"
+    SERVER_NAME = os.environ.get("SERVER_NAME")
 
     # Email sending
     MAIL_SERVER = "smtp.gmail.com"
