+++ happiness-frontend/src/components/happinessHistory/Histories.js
