+++ happiness-frontend/src/components/happinessHistory/MonthView.js
