+++ happiness-frontend/src/components/statGraphs/ChartPreview.js
