+++ happiness-frontend/src/components/statGraphs/Graph.js
