+++ happiness-frontend/src/components/statGraphs/LineChart.js
