+++ happiness-frontend/src/components/statGraphs/Stat.js
