+++ happiness-frontend/src/components/submitHappiness/DynamicSmile.js
